pyproject.toml
[project]
name = "lazy-linux"
version = "0.1.0"
description = "Awesome smart OS"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click (>=8.2.1,<9.0.0)",
    "rich (>=14.0.0,<15.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "openai (>=1.93.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
lazy-linux = "lazy_linux:cli"
